+++ src/tools/web-browsing/Render/Search/index.tsx
@@ -3,7 +3,7 @@ import { memo, useState } from 'react';
 import { Flexbox } from 'react-layout-kit';
 
 import { ChatMessagePluginError } from '@/types/message';
-import { SearchQuery, SearchResponse } from '@/types/tool/search';
+import { SearchQuery, UniformSearchResponse } from '@/types/tool/search';
 
 import ConfigForm from './ConfigForm';
 import SearchQueryView from './SearchQuery';
@@ -13,7 +13,7 @@ interface SearchProps {
   messageId: string;
   pluginError: ChatMessagePluginError;
   searchQuery: SearchQuery;
-  searchResponse?: SearchResponse;
+  searchResponse?: UniformSearchResponse;
 }
 
 const Search = memo<SearchProps>(({ messageId, searchQuery, searchResponse, pluginError }) => {